Fix "Uncaught TypeError: Cannot read property 'props' of undefined" when setDocumentTitle={true} and excludes={["App"]}
There is a bug which appears on the root route, when breadcrumb is configured with `setDocumentTitle={true}` and `excludes={["App"]}`. The bug is due to a missing check for array length in `crumbs[crumbs.length-1].props`, so when `crumbs = []` it tries to access `crumbs[-1]` which obviously fails.

Please merge and publish update.