chore(demo): rewrite/redesign demo
Start updating the demo to work with the latest version of
the package. `npm init`ed the `/demo` directory so it can
contain it's own `package.json` and `dependencies`. Also
added `webpack-dev-server` for an easier dev workflow and
started to refactor the webpack configurations. The demo
still needs a little more work but it's getting close.