fix(demo): update crumb-route
Update `CrumbRoute` to use `match.path` from `routeProps` instead of
the actual `path` passed to a route. This prevents issues with parameters
(e.g. `/people/:id`) and will still have the breadcrumb navigate to the route
that corresponds to that path. I also added an `includeSearch` prop which
allows users to keep any search parameters in the breadcrumb link.